+++ Mail/src/main/java/uns/ac/rs/Mail/service/MessageServiceInterface.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface MessageServiceInterface {
 
-    Message findOne(Integer msgId);
+    Message findOne(Long msgId);
     List<Message> findAll();
     List<Message> findAllOrderByDate();
